requirements.txt
asgiref==3.8.1
Django==3.2
django-cors-headers==4.3.1
djangorestframework==3.15.1
pytz==2024.1
setuptools==70.0.0
sqlparse==0.5.0
tzdata==2024.1
